AQ-56: refactor folders
