+++ app/src/main/java/com/example/monitoreoacua/service/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.example.monitoreoacua.models.request;
+package com.example.monitoreoacua.service.request;
 
 public class LoginRequest extends BaseRequest {
