+++ app/src/main/java/com/example/monitoreoacua/views/menu/HomeActivity.java
@@ -1,21 +1,17 @@
 package com.example.monitoreoacua.views.menu;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.monitoreoacua.R;
-import com.example.monitoreoacua.models.request.LogoutRequest;
-import com.example.monitoreoacua.models.response.ApiResponse;
-import com.example.monitoreoacua.models.response.LogoutResponse;
+import com.example.monitoreoacua.service.request.LogoutRequest;
+import com.example.monitoreoacua.service.response.LogoutResponse;
 import com.example.monitoreoacua.service.ApiClient;
 import com.example.monitoreoacua.service.ApiUsersService;
 import com.example.monitoreoacua.views.login.LoginActivity;
